Dotty 3.0.0-M1 support done. Remove structural types
- Action.Trigger.Wrapper, Oriented.Wrapper, Orientable.Wrapper have been
  removed; one must mix in Action.Trigger, Oriented, Orientable instead
  and implement the methods
- Swing.PeerContainer is not a structural type but a trait, that must
  be mixed in
- clean up some examples
- remove lanuage feature lazy vals from package object

+++ examples/src/main/scala/scala/swing/examples/LinePainting.scala
@@ -47,11 +47,11 @@ object LinePainting extends SimpleSwingApplication {
     var path = new geom.GeneralPath
 
     def lineTo(p: Point): Unit = {
-      path.lineTo(p.x, p.y); repaint()
+      path.lineTo(p.x.toFloat, p.y.toFloat); repaint()
     }
 
     def moveTo(p: Point): Unit = {
-      path.moveTo(p.x, p.y); repaint()
+      path.moveTo(p.x.toFloat, p.y.toFloat); repaint()
     }
 
     override def paintComponent(g: Graphics2D): Unit = {
